# python-sdk/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@katanemo.com
    name: Katanemo Labs Inc.
    url: https://api.katanemo.com
  description: "With Katanemo developers can add support for users, enterprise SSO,\
    \ machine keys and fine-grained authorization in minutes."
  termsOfService: https://www.katanemo.com/terms-of-service
  title: "Katanemo - identity, and fine-grained authorization for modern { API-first\
    \ } software companies."
  version: 1.0.0
servers:
- url: https://api.katanemo.com
...
paths:
  /service:
    get:
      description: List services that belong to a particular developer. Requires bearer
        token authorization
      operationId: ListServicesByDeveloper
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ServiceResponseObj'
                type: array
          description: List all services belonging to a developer
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: List Services
      tags:
      - service
    post:
      description: Create a Service in Katanemo. Once a service is created Katanemo  identity
        and authorization capabilities on behalf of SaaS (API) Developers
      operationId: createService
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
        description: "Details about the service, including the text-representation\
          \ of the service APIs."
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponseObj'
          description: Service Successfully Created.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Create Service
      tags:
      - service
  /service/3xA:
    get:
      description: "Gets details about Katanemo's AAA SaaS service. Katanemo is powered\
        \ by Katanemo, and our 3xA service uses the same core identity and authorization\
        \ capabilities that we offer SaaS (API) developers"
      operationId: GetDefaultService
      responses:
        "200":
          content:
            application/json; charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ServiceResponseObj'
          description: service
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get Details for Katanemo AAA
      tags:
      - service
  /service/{serviceId}:
    delete:
      description: Deletes a service. Note the delete operation is a 'soft' delete
        where by organizations can't access your service. Requires a bearer token
        to validate that the caller can delete the service.
      operationId: deleteService
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Deleted service.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Delete Service
      tags:
      - service
    get:
      description: "Gets a Katanemo Service. The principal token must be present in\
        \ the bearer header to retrieve the service details, unless the service is\
        \ public"
      operationId: getService
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponseObj'
          description: service
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Get Service
      tags:
      - service
    put:
      description: Update Service. Requires bearer token authorization for the caller
        updating the service
      operationId: updateService
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
        description: Service object of the service being updated.
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceResponseObj'
          description: Updated service object.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Update Service
      tags:
      - service
  /service/{serviceId}/tags:
    get:
      description: Gets tags associated with a resource of a service
      operationId: GetTagsForResource
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetTagsRequest'
        description: Tags and resource id
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tags'
          description: tags
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
      summary: Gets tags for a resource
      tags:
      - service
      - access-control
    post:
      description: "Add tags (key/value pair) to a particular resource that is created\
        \ for a service, for a particular organization account id."
      operationId: createTags
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tags'
        description: Tags and resource id
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tags'
          description: User account
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: Add tags to a resource
      tags:
      - service
      - access-control
  /service/{serviceId}/.well-known/jwks.json:
    get:
      description: Gets public key that can be used to verify jwt token issued by
        Katanemo. This API does not require bearer authorization
      operationId: GetDeveloperPublicKeys
      parameters:
      - explode: false
        in: path
        name: serviceId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDeveloperPublicKeys_200_response'
          description: List of public keys for developer
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestException'
          description: Bad Request Exception
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedException'
          description: Unauthorized Exception
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictException'
          description: Conflict Exception
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TooManyRequestsException'
          description: Too Many Requests Exception
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerErrorException'
          description: Internal Server Error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      summary: JWKS endpoint for Service
      tags:
      - service
...
components:
  schemas:
...
    CreateServiceRequest:
      properties:
        name:
          description: Service Name
          type: string
        description:
          description: Service Description
          type: string
        redirectUrl:
          description: Redirect URL after a successful login.
          type: string
        apiSpecFile:
          description: openapi service json or yaml file
          format: binary
          type: string
        authExclusionPaths:
          description: List of paths for which we do not require authentication
          items:
            type: string
          type: array
        displayName:
          description: "Display name of the service/company used in the Sign up, Login,\
            \ Logout and other relevant branding pages"
          type: string
        logoUrl:
          description: "The URL for the service/company Logo used in the Sign up,\
            \ Login, Logout and other relevant branding pages"
          type: string
        detailsImageUrl:
          description: The URL of image showing details about the service to be displayed
            on the sign-up page.
          type: string
        termsUrl:
          description: The URL for the terms of the service
          type: string
        privacyUrl:
          description: The URL for the privacy of the service
          type: string
        docsUrl:
          description: The URL for the documentation of the service
          type: string
      required:
      - apiSpecFile
      - name
      - redirectUrl
      type: object
...
    ServiceResponseObj:
      example:
        privacyUrl: privacyUrl
        apiSpecFileContents: apiSpecFileContents
        redirectUrl: redirectUrl
        authExclusionPaths:
        - authExclusionPaths
        - authExclusionPaths
        displayName: displayName
        description: description
        onboardURL: onboardURL
        serviceName: serviceName
        version: 0
        logoUrl: logoUrl
        accountId: accountId
        createdAt: 1
        termsUrl: termsUrl
        apis:
        - apis
        - apis
        docsUrl: docsUrl
        serviceId: serviceId
        updatedAt: 6
        detailsImageUrl: detailsImageUrl
      properties:
        accountId:
          description: Account ID of service creator
          title: accountId
          type: string
        serviceId:
          description: ID of the service
          title: serviceId
          type: string
        serviceName:
          description: Name of the service
          title: serviceName
          type: string
        description:
          description: Description of service
          title: description
          type: string
        onboardURL:
          description: Pretty onboarding URL for the service generated by Katanemo
          title: onboardURL
          type: string
        redirectUrl:
          description: redirect url
          title: redirectUrl
          type: string
        apiSpecFileContents:
          description: S3 url of open api spec (Json or yaml) file contents
          title: apiSpecFileContents
          type: string
        apis:
          description: List of paths exposed by the service
          items:
            type: string
          title: apis
          type: array
        authExclusionPaths:
          description: List of paths for which we do not require authentication
          items:
            type: string
          title: authExclusionPaths
          type: array
        version:
          title: version
          type: integer
        updatedAt:
          format: int64
          title: updatedAt
          type: integer
        createdAt:
          format: int64
          title: createdAt
          type: integer
        displayName:
          description: "Display name of the service/company used in the Sign up, Login,\
            \ Logout and other relevant branding pages"
          title: displayName
          type: string
        logoUrl:
          description: "The URL for the service/company Logo used in the Sign up,\
            \ Login, Logout and other relevant branding pages"
          title: logoUrl
          type: string
        detailsImageUrl:
          description: The URL of image showing details about the service to be displayed
            on the sign-up page.
          title: detailsImageUrl
          type: string
        termsUrl:
          description: The URL for the terms of the service
          title: termsUrl
          type: string
        privacyUrl:
          description: The URL for the privacy of the service
          title: privacyUrl
          type: string
        docsUrl:
          description: The URL for the documentatio of the service
          title: docsUrl
          type: string
      required:
      - apiSpecFileContents
      - apis
      - onboardURL
      - redirectUrl
      - serviceId
      - serviceName
      title: ServiceResponseObj
...
    Error:
      example:
        code: 0
        message: message
      properties:
        code:
          description: Error code
          format: int32
          title: code
          type: integer
        message:
          description: Error message
          title: message
          type: string
      required:
      - code
      - message
      title: Error
    BadRequestException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: BadRequestException
      type: object
    UnauthorizedException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: UnauthorizedException
      type: object
...
    ConflictException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: ConflictException
      type: object
    TooManyRequestsException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: TooManyRequestsException
      type: object
    InternalServerErrorException:
      properties:
        message:
          title: message
          type: string
        errorCode:
          title: errorCode
          type: string
      title: InternalServerErrorException
      type: object
...